requirements.txt
# Core dependencies for Regime-Aware Stock Forecasting
numpy>=1.24.0
pandas>=2.0.0
tensorflow-cpu>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.2
scikit-learn>=1.3.0
scipy>=1.11.0

# Technical analysis
ta>=0.10.0

# Data sources
yfinance>=0.1.70
fredapi>=0.5.0

# Environment management
python-dotenv>=0.19.0

# Optional for improved performance
numba>=0.56.0



